+++ test/Utils/Assertion/AssertsPattern.php
@@ -14,6 +14,11 @@ trait AssertsPattern
         $this->assertTrue($pattern->test($subject), "Failed to assert that $pattern matches '$subject'");
     }
 
+    private function assertPatternFails(Pattern $pattern, string $subject): void
+    {
+        $this->assertTrue($pattern->fails($subject), "Failed to assert that $pattern fails to match '$subject'");
+    }
+
     public function assertPatternIs(string $expected, Pattern $actual): void
     {
         Assert::assertSame($expected, $actual->delimited());
@@ -21,12 +26,17 @@ trait AssertsPattern
 
     public function assertConsumesFirst(string $text, Pattern $pattern): void
     {
-        $this->assertSame($pattern->match($text)->first()->text(), $text);
+        $this->assertSame($text, $pattern->search($text)->first());
+    }
+
+    public function assertConsumesFirstGroup(string $subject, string $expectedGroup, Pattern $pattern): void
+    {
+        $this->assertSame($expectedGroup, $pattern->match($subject)->first()->group(1)->text());
     }
 
     public function assertConsumesAll(string $text, array $texts, Pattern $pattern): void
     {
-        $this->assertSame($pattern->search($text)->all(), $texts);
+        $this->assertSame($texts, $pattern->search($text)->all());
     }
 
     private function assertGroupMissing(Matcher $matcher, int $nameOrIndex): void